Move ContentItem search up into main data-source
